### 2024-11-15 #7488 @editorialbot
[PRE REVIEW]: Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Planetary Defense
**Submitting author:** @rahil-makadia (<a href="http://orcid.org/0000-0001-9265-2230">Rahil Makadia</a>)
**Repository:** https://github.com/rahil-makadia/grss
**Branch with paper.md** (empty if default branch): dev
**Version:** v4.3.6
