### 2024-11-15 #7488 @warrickball
For whichever editor might pick this up, note that this is a joint publication with AAS Journals (specifically, the *Planetary Science Journal*). :earth_africa:
