### 2024-11-15 #7488 @rahil-makadia
@warrickball Looks like it might take some time before the review starts on this paper. Let me know if there's anything to do in the meantime or if it's just a waiting game right now...
