### 2024-11-15 #7488 @warrickball
I'm afraid it is mostly a waiting game. Hopefully an editor will become available within a week or two, and they'll then start seeking reviewers.

The basic metadata properties all look good. My only comment on the paper at this point is that [we prefer inline hyperlinks to footnotes](https://joss.readthedocs.io/en/latest/paper.html#links).
