+++ docs/source/prop.rst
@@ -9,7 +9,7 @@ The force model includes the following effects:
 #. J\ :sub:`2` zonal harmonic effects from the Sun and the Earth
 #. A\ :sub:`1`, A\ :sub:`2`, A\ :sub:`3` nongravitational  acceleration model [#]_
 
-In addition to any integrated bodies defined by the user, the default force model includes effects from the Sun, the planets, the Moon, Pluto, and the Big 16 main-belt asteroids. The states of these bodies are read in from either the JPL DE431 or DE441 ephemerides, based on the user's choice. The propagator can also interpolate any of the integrated bodies' states to any time within the integration (or within a margin around the integration time bounds). This interpolation is done via interpolation polynomials based on the Newton divided difference method.
+In addition to any integrated bodies defined by the user, the default force model includes effects from the Sun, the planets, the Moon, Pluto, and the Big 16 main-belt asteroids. The states of these bodies are read in from either the JPL DE431 or DE441 ephemerides, based on the user's choice. The propagator can also interpolate any of the integrated bodies' states to any time within the integration (or within a margin around the integration time bounds). This interpolation is done directly via polynomial coefficients computed by IAS15 during propagation.
 
 The propagator module also has the ability to calculate apparent states of any integrated body. This is done through two methods. The first is the simpler of the two, and uses a simple approximation of the light-time between an observer and the integrated body. The second method is more accurate, and uses an iterative method to calculate the light-time between an observer and the integrated body. The second method is slightly more computationally expensive since it requires a few extra interpolations.
 
