### 2024-11-21 3590cb3
changed paper footnotes to hyperlinks
