### 2024-11-21 #7488 @rahil-makadia
I have updated the footnotes to inline hyperlinks as per suggestion
