### 2024-12-01 ec2e2ad
only save machine data for parallel propagations
