+++ joss/paper.md
@@ -26,7 +26,6 @@ affiliations:
     index: 1
   - name: Jet Propulsion Laboratory, California Institute of Technology, Pasadena, CA 91109, USA
     index: 2
-date: XX August 2024
 bibliography: paper.bib
 aas-doi: 10.3847/PSJ/xxxxx
 aas-journal: Planetary Science Journal
