+++ src/simulation.cpp
@@ -1371,8 +1371,9 @@ void PropSimulation::extend(real tf, std::vector<real> tEvalNew,
 
 /**
  * @param[in] filename Name of the file to save the simulation to.
+ * @param[in] onlyMachineData Flag to save only machine-readable data (default: false).
  */
-void PropSimulation::save(std::string filename) {
+void PropSimulation::save(std::string filename, bool onlyMachineData) {
     struct stat fileExistsBuffer;
     if (stat(filename.c_str(), &fileExistsBuffer) == 0) {
         throw std::invalid_argument("Cannot save PropSimulation '" + this->name +
@@ -1413,6 +1414,10 @@ void PropSimulation::save(std::string filename) {
     std::string nextSubsection = this->name;
     file << std::string((int)(maxChars-nextSubsection.size())/2, '-') << nextSubsection << std::string((int)(maxChars-nextSubsection.size())/2, '-') << std::endl;
     file << subsectionFull << std::endl;
+    size_t starti = 0;
+    if (onlyMachineData) {
+        goto machineData;
+    }
     file << "Integration from MJD " << timeWidth << std::fixed << timeFloatPrec << this->integParams.t0
          << " to MJD " << timeWidth << std::fixed << timeFloatPrec << this->integParams.tf << " [TDB]"
          << std::endl;
@@ -1423,7 +1428,7 @@ void PropSimulation::save(std::string filename) {
     nextSubsection = std::to_string(this->integParams.nInteg) + " Integration bodies";
     file << std::string((int)(maxChars-nextSubsection.size())/2, '-') << nextSubsection << std::string((int)(maxChars-nextSubsection.size())/2, '-') << std::endl;
     file << subsectionFull << std::endl;
-    size_t starti = 0;
+    starti = 0;
     for (size_t i = 0; i < this->integBodies.size(); i++) {
         file << this->integBodies[i].name << std::endl;
         file << "Initial time : MJD " << timeWidth << std::fixed << timeFloatPrec << this->integBodies[i].t0
@@ -1541,6 +1546,7 @@ void PropSimulation::save(std::string filename) {
     }
 
     // insert machine readable data for close approach and impact parameters
+    machineData : {
     file.precision(18);
     if (this->impactParams.size() > 0) {
         file << std::endl;
@@ -1700,6 +1706,7 @@ void PropSimulation::save(std::string filename) {
         }
         file << "$$CA_END" << std::endl;
     }
+    } // end of machine data
 
     file << std::endl;
     file << sectionFull << std::endl;
