### 2024-12-20 ecd69b0
copyright and credit notices; gr15 -> ias15
