+++ src/spk.cpp
@@ -1,3 +1,33 @@
+/**
+ * @file    spk.cpp
+ * @brief   Source file for the SPK ephemeris data.
+ * @author  Rahil Makadia <makadia2@illinois.edu>
+ * @details This file implements the routines to read in the binary SPK files
+ * provided by NAIF and to interpolate the ephemeris data to a requested time.
+ * The routines here are based on the small-body ephemeris reading routines
+ * implemented by Holman et al. (2023) in the ASSIST library. However, those
+ * rountines are extended here to also handle the planetary ephemeris data with
+ * the same code rather than separatley using different ephemeris file types.
+ *
+ * @section     LICENSE
+ * Copyright (C) 2022-2025 Rahil Makadia
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; either version 3 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program; if not, see <https://www.gnu.org/licenses>.
+ */
+
+// code here is heavily modified but taken from
+// https://github.com/matthewholman/assist/tree/main/src/spk
 #include "spk.h"
 
 /**
