+++ _sources/cppsummary/gr15.rst
@@ -1,6 +1,0 @@
-gr15.h
-======
-
-.. doxygenfile:: gr15.h
-   :project: GRSS
-
