+++ grss/debias/get_debiasing_data.py
@@ -8,12 +8,12 @@ FTP_SITE = 'ftp://ssd.jpl.nasa.gov/pub/ssd/debias'
 # if lowres_data.tgz or lowres_data/ already exist, do not download
 if not os.path.exists(f'{cwd}/lowres_data.tgz') and not os.path.exists(f'{cwd}/lowres_data/'):
     FTP_FILE = 'debias_2018.tgz'
-    cmd = f'curl --silent --show-error -o {cwd}/lowres_data.tgz {FTP_SITE}/{FTP_FILE}'
+    cmd = f'curl --connect-timeout 30 --silent --show-error -o {cwd}/lowres_data.tgz {FTP_SITE}/{FTP_FILE}'
     print('Downloading low-resolution debiasing data...')
     os.system(cmd)
 if not os.path.exists(f'{cwd}/hires_data.tgz') and not os.path.exists(f'{cwd}/hires_data/'):
     FTP_FILE = 'debias_hires2018.tgz'
-    cmd = f'curl --silent --show-error -o {cwd}/hires_data.tgz {FTP_SITE}/{FTP_FILE}'
+    cmd = f'curl --connect-timeout 30 --silent --show-error -o {cwd}/hires_data.tgz {FTP_SITE}/{FTP_FILE}'
     print('Downloading high-resolution debiasing data...')
     os.system(cmd)
 
