### 2025-01-15 #7488 @warrickball
I'm afraid not, sorry!. We're still just waiting for an editor to pick this up and we have a substantial backlog. Hopefully it won't hold up the AAS publication but let us know when if the paper moves to accepted and I'll prioritise this on our waiting list.
