+++ joss/paper.bib
@@ -26,12 +26,12 @@ doi       = {10.3847/PSJ/ad1bce},
 publisher = {The American Astronomical Society},
 }
 
-@article{Makadia2024b,
+@article{Makadia2025,
 author    = {Rahil Makadia and Davide Farnocchia and Steven R. Chesley and Siegfried Eggl},
 journal   = {The Planetary Science Journal},
 title     = {{Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Planetary Defense}},
-year      = {2024},
-volume    = {Submitted},
+year      = {2025},
+volume    = {Accepted},
 publisher = {The American Astronomical Society},
 }
 
