### 2025-02-18 #7488 @warrickball
> @warrickball the PSJ paper was just accepted this morning. I've updated the citation in the JOSS manuscript. If it helps with the review, I can also put a link to the accepted PSJ manuscript here...

Thanks, though it isn't necessary. If you have a DOI (even if it doesn't resolve yet), you can update the `aas-doi` field in the paper's YAML metadata.

I'm still trying to find a second reviewer, which I'm hoping will take less than a week. Sorry for the continued wait!
