### 2025-02-28 #7488 @warrickball
Hi @athulpg007, would you possibly be interested in reviewing this submission for JOSS? I see you've reviewed similar submissions before (and thank you for that! 🙏). If you aren't able to review, I'd be grateful for any suggestions, either here or by email. Thanks!
