### 2025-02-28 #7861 @editorialbot
[REVIEW]: Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Planetary Defense
**Submitting author:** @rahil-makadia (<a href="http://orcid.org/0000-0001-9265-2230">Rahil Makadia</a>)
**Repository:** https://github.com/rahil-makadia/grss
**Branch with paper.md** (empty if default branch): main
**Version:** v4.4.0
