### 2025-03-05 ac6b2ba
add psj paper doi to joss draft
