### 2025-03-06 #7861 @warrickball
Sorry for missing this. The `@editorialbot` command has to be the first line of the message.
