### 2025-03-13 rahil-makadia/grss#89 @athulpg007
Installation issue via source code (Python)
I was able to install grss from PyPI and run the example notebooks without issues. However, when following the instructions here[: Install via source code (Python) ](https://rahil-makadia.github.io/grss/start.html#install-via-source-code-python) the installation fails. It does not impair the usage of the package, and may be specific to my machine. But if it's a real bug, please consider fixing.

There's 2 errors:
```
    Traceback (most recent call last):
        File "/home/athul/Documents/grss/.venv/bin/pybind11-config", line 5, in <module>
          from pybind11.__main__ import main
      ModuleNotFoundError: No module named 'pybind11'
```
though pybind11 is installed, and
```
 Consolidate compiler generated dependencies of target apophis
      make[2]: *** No rule to make target '../extern/cspice/lib/cspice.a', needed by '../tests/cpp/prop/apophis.out'.  Stop.
      make[1]: *** [CMakeFiles/Makefile2:176: tests/cpp/prop/CMakeFiles/apophis.dir/all] Error 2
      make: *** [Makefile:91: all] Error 2

```

The full stacktrace is attached below.

Steps to reproduce:
```
$ git clone https://www.github.com/rahil-makadia/grss
cd grss
source initialize.sh
<created a Python 3.12 venv>
python3 -m pip install .
```
Result:
```
(.venv) athul@athul-Lenovo-YOGA-720-13IKB:~/Documents/grss$ python3 -m pip install .
Processing /home/athul/Documents/grss
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting astropy>=5.0 (from grss==4.3.8)
  Obtaining dependency information for astropy>=5.0 from https://files.pythonhosted.org/packages/ed/70/930a24bd3f2046ba2c4b8a914d0349ffb17760d044f2a4e8c3195b984851/astropy-7.0.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached astropy-7.0.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (10 kB)
Collecting astroquery>=0.4.0 (from grss==4.3.8)
  Obtaining dependency information for astroquery>=0.4.0 from https://files.pythonhosted.org/packages/fb/e5/ec215258bdc718a9a1ccfd44332af9824e9a4be60db96afce449dac81f79/astroquery-0.4.9.post1-py3-none-any.whl.metadata
  Using cached astroquery-0.4.9.post1-py3-none-any.whl.metadata (6.3 kB)
Collecting basemap>=1.4.0 (from grss==4.3.8)
  Obtaining dependency information for basemap>=1.4.0 from https://files.pythonhosted.org/packages/d7/ee/5908870dab84ea5c3f3f29b665a6042d8a461cfd315c701c02d10d607ef0/basemap-1.4.1-cp312-cp312-manylinux1_x86_64.whl.metadata
  Using cached basemap-1.4.1-cp312-cp312-manylinux1_x86_64.whl.metadata (9.1 kB)
Collecting healpy>=1.16.0 (from grss==4.3.8)
  Obtaining dependency information for healpy>=1.16.0 from https://files.pythonhosted.org/packages/04/92/989ce55bdb99d09ec40199830984e7746c17362a6637733c75bac6675453/healpy-1.18.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached healpy-1.18.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (8.7 kB)
Collecting lxml>=5.1.0 (from grss==4.3.8)
  Obtaining dependency information for lxml>=5.1.0 from https://files.pythonhosted.org/packages/c5/f8/309546aec092434166a6e11c7dcecb5c2d0a787c18c072d61e18da9eba57/lxml-5.3.1-cp312-cp312-manylinux_2_28_x86_64.whl.metadata
  Using cached lxml-5.3.1-cp312-cp312-manylinux_2_28_x86_64.whl.metadata (3.7 kB)
Collecting matplotlib>=3.3.0 (from grss==4.3.8)
  Obtaining dependency information for matplotlib>=3.3.0 from https://files.pythonhosted.org/packages/63/23/c0615001f67ce7c96b3051d856baedc0c818a2ed84570b9bf9bde200f85d/matplotlib-3.10.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached matplotlib-3.10.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (11 kB)
Collecting numba>=0.55.0 (from grss==4.3.8)
  Obtaining dependency information for numba>=0.55.0 from https://files.pythonhosted.org/packages/18/74/6a9f0e6c76c088f8a6aa702eab31734068061dca5cc0f34e8bc1eb447de1/numba-0.61.0-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata
  Using cached numba-0.61.0-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (2.8 kB)
Collecting numpy>=1.22.0 (from grss==4.3.8)
  Obtaining dependency information for numpy>=1.22.0 from https://files.pythonhosted.org/packages/39/04/78d2e7402fb479d893953fb78fa7045f7deb635ec095b6b4f0260223091a/numpy-2.2.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached numpy-2.2.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (62 kB)
Collecting pandas>=1.5.0 (from grss==4.3.8)
  Obtaining dependency information for pandas>=1.5.0 from https://files.pythonhosted.org/packages/38/f8/d8fddee9ed0d0c0f4a2132c1dfcf0e3e53265055da8df952a53e7eaf178c/pandas-2.2.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached pandas-2.2.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (89 kB)
Requirement already satisfied: pybind11[global]>=2.10.0 in ./.venv/lib/python3.12/site-packages (from grss==4.3.8) (2.13.6)
Requirement already satisfied: requests>=2.22.0 in ./.venv/lib/python3.12/site-packages (from grss==4.3.8) (2.32.3)
Collecting tqdm>=4.65.0 (from grss==4.3.8)
  Obtaining dependency information for tqdm>=4.65.0 from https://files.pythonhosted.org/packages/d0/30/dc54f88dd4a2b5dc8a0279bdd7270e735851848b762aeb1c1184ed1f6b14/tqdm-4.67.1-py3-none-any.whl.metadata
  Using cached tqdm-4.67.1-py3-none-any.whl.metadata (57 kB)
Collecting pyerfa>=2.0.1.1 (from astropy>=5.0->grss==4.3.8)
  Obtaining dependency information for pyerfa>=2.0.1.1 from https://files.pythonhosted.org/packages/e5/e0/050018d855d26d3c0b4a7d1b2ed692be758ce276d8289e2a2b44ba1014a5/pyerfa-2.0.1.5-cp39-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached pyerfa-2.0.1.5-cp39-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.7 kB)
Collecting astropy-iers-data>=0.2025.1.31.12.41.4 (from astropy>=5.0->grss==4.3.8)
  Obtaining dependency information for astropy-iers-data>=0.2025.1.31.12.41.4 from https://files.pythonhosted.org/packages/c9/a1/c4d0b1f17be69a1eb4d50bc98f465d98cade3a1277d0d6c933f8a01af528/astropy_iers_data-0.2025.3.10.0.29.26-py3-none-any.whl.metadata
  Using cached astropy_iers_data-0.2025.3.10.0.29.26-py3-none-any.whl.metadata (5.1 kB)
Requirement already satisfied: PyYAML>=6.0.0 in ./.venv/lib/python3.12/site-packages (from astropy>=5.0->grss==4.3.8) (6.0.2)
Requirement already satisfied: packaging>=22.0.0 in ./.venv/lib/python3.12/site-packages (from astropy>=5.0->grss==4.3.8) (24.2)
Requirement already satisfied: beautifulsoup4>=4.8 in ./.venv/lib/python3.12/site-packages (from astroquery>=0.4.0->grss==4.3.8) (4.13.3)
Collecting html5lib>=0.999 (from astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for html5lib>=0.999 from https://files.pythonhosted.org/packages/6c/dd/a834df6482147d48e225a49515aabc28974ad5a4ca3215c18a882565b028/html5lib-1.1-py2.py3-none-any.whl.metadata
  Using cached html5lib-1.1-py2.py3-none-any.whl.metadata (16 kB)
Collecting keyring>=15.0 (from astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for keyring>=15.0 from https://files.pythonhosted.org/packages/d3/32/da7f44bcb1105d3e88a0b74ebdca50c59121d2ddf71c9e34ba47df7f3a56/keyring-25.6.0-py3-none-any.whl.metadata
  Using cached keyring-25.6.0-py3-none-any.whl.metadata (20 kB)
Collecting pyvo>=1.5 (from astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for pyvo>=1.5 from https://files.pythonhosted.org/packages/64/0c/7b56358a1722d409914749f42b4e398233352a3cca365b280e3ca96ea2ff/pyvo-1.6.1-py3-none-any.whl.metadata
  Using cached pyvo-1.6.1-py3-none-any.whl.metadata (4.8 kB)
Collecting basemap-data<1.4,>=1.3.2 (from basemap>=1.4.0->grss==4.3.8)
  Obtaining dependency information for basemap-data<1.4,>=1.3.2 from https://files.pythonhosted.org/packages/2c/c3/63fdb885308c999206c7a497d79a50891581626da7dc491d432d2732bb46/basemap_data-1.3.2-py2.py3-none-any.whl.metadata
  Using cached basemap_data-1.3.2-py2.py3-none-any.whl.metadata (2.7 kB)
Collecting pyshp<2.4,>=1.2 (from basemap>=1.4.0->grss==4.3.8)
  Obtaining dependency information for pyshp<2.4,>=1.2 from https://files.pythonhosted.org/packages/98/2f/68116db5b36b895c0450e3072b8cb6c2fac0359279b182ea97014d3c8ac0/pyshp-2.3.1-py2.py3-none-any.whl.metadata
  Using cached pyshp-2.3.1-py2.py3-none-any.whl.metadata (55 kB)
Collecting matplotlib>=3.3.0 (from grss==4.3.8)
  Obtaining dependency information for matplotlib>=3.3.0 from https://files.pythonhosted.org/packages/cd/ba/412149958e951876096198609b958b90a8a2c9bc07a96eeeaa9e2c480f30/matplotlib-3.8.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached matplotlib-3.8.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.8 kB)
Collecting pyproj<3.7.0,>=1.9.3 (from basemap>=1.4.0->grss==4.3.8)
  Obtaining dependency information for pyproj<3.7.0,>=1.9.3 from https://files.pythonhosted.org/packages/19/9b/c57569132174786aa3f72275ac306956859a639dad0ce8d95c8411ce8209/pyproj-3.6.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached pyproj-3.6.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (31 kB)
Collecting packaging>=22.0.0 (from astropy>=5.0->grss==4.3.8)
  Obtaining dependency information for packaging>=22.0.0 from https://files.pythonhosted.org/packages/ec/1a/610693ac4ee14fcdf2d9bf3c493370e4f2ef7ae2e19217d7a237ff42367d/packaging-23.2-py3-none-any.whl.metadata
  Using cached packaging-23.2-py3-none-any.whl.metadata (3.2 kB)
Collecting numpy>=1.22.0 (from grss==4.3.8)
  Obtaining dependency information for numpy>=1.22.0 from https://files.pythonhosted.org/packages/0f/50/de23fde84e45f5c4fda2488c759b69990fd4512387a8632860f3ac9cd225/numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (61 kB)
Collecting contourpy>=1.0.1 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for contourpy>=1.0.1 from https://files.pythonhosted.org/packages/ba/99/6794142b90b853a9155316c8f470d2e4821fe6f086b03e372aca848227dd/contourpy-1.3.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached contourpy-1.3.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.4 kB)
Collecting cycler>=0.10 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for cycler>=0.10 from https://files.pythonhosted.org/packages/e7/05/c19819d5e3d95294a6f5947fb9b9629efb316b96de511b418c53d245aae6/cycler-0.12.1-py3-none-any.whl.metadata
  Using cached cycler-0.12.1-py3-none-any.whl.metadata (3.8 kB)
Collecting fonttools>=4.22.0 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for fonttools>=4.22.0 from https://files.pythonhosted.org/packages/95/70/2a781bedc1c45a0c61d29c56425609b22ed7f971da5d7e5df2679488741b/fonttools-4.56.0-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached fonttools-4.56.0-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (101 kB)
Collecting kiwisolver>=1.3.1 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for kiwisolver>=1.3.1 from https://files.pythonhosted.org/packages/bc/b3/9458adb9472e61a998c8c4d95cfdfec91c73c53a375b30b1428310f923e4/kiwisolver-1.4.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached kiwisolver-1.4.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (6.2 kB)
Collecting pillow>=8 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for pillow>=8 from https://files.pythonhosted.org/packages/38/0d/84200ed6a871ce386ddc82904bfadc0c6b28b0c0ec78176871a4679e40b3/pillow-11.1.0-cp312-cp312-manylinux_2_28_x86_64.whl.metadata
  Using cached pillow-11.1.0-cp312-cp312-manylinux_2_28_x86_64.whl.metadata (9.1 kB)
Collecting pyparsing>=2.3.1 (from matplotlib>=3.3.0->grss==4.3.8)
  Obtaining dependency information for pyparsing>=2.3.1 from https://files.pythonhosted.org/packages/1c/a7/c8a2d361bf89c0d9577c934ebb7421b25dc84bf3a8e3ac0a40aed9acc547/pyparsing-3.2.1-py3-none-any.whl.metadata
  Using cached pyparsing-3.2.1-py3-none-any.whl.metadata (5.0 kB)
Requirement already satisfied: python-dateutil>=2.7 in ./.venv/lib/python3.12/site-packages (from matplotlib>=3.3.0->grss==4.3.8) (2.9.0.post0)
Collecting llvmlite<0.45,>=0.44.0dev0 (from numba>=0.55.0->grss==4.3.8)
  Obtaining dependency information for llvmlite<0.45,>=0.44.0dev0 from https://files.pythonhosted.org/packages/cb/da/8341fd3056419441286c8e26bf436923021005ece0bff5f41906476ae514/llvmlite-0.44.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata
  Using cached llvmlite-0.44.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.0 kB)
Collecting pytz>=2020.1 (from pandas>=1.5.0->grss==4.3.8)
  Obtaining dependency information for pytz>=2020.1 from https://files.pythonhosted.org/packages/eb/38/ac33370d784287baa1c3d538978b5e2ea064d4c1b93ffbd12826c190dd10/pytz-2025.1-py2.py3-none-any.whl.metadata
  Using cached pytz-2025.1-py2.py3-none-any.whl.metadata (22 kB)
Collecting tzdata>=2022.7 (from pandas>=1.5.0->grss==4.3.8)
  Obtaining dependency information for tzdata>=2022.7 from https://files.pythonhosted.org/packages/0f/dd/84f10e23edd882c6f968c21c2434fe67bd4a528967067515feca9e611e5e/tzdata-2025.1-py2.py3-none-any.whl.metadata
  Using cached tzdata-2025.1-py2.py3-none-any.whl.metadata (1.4 kB)
Requirement already satisfied: pybind11-global==2.13.6 in ./.venv/lib/python3.12/site-packages (from pybind11[global]>=2.10.0->grss==4.3.8) (2.13.6)
Requirement already satisfied: charset-normalizer<4,>=2 in ./.venv/lib/python3.12/site-packages (from requests>=2.22.0->grss==4.3.8) (3.4.1)
Requirement already satisfied: idna<4,>=2.5 in ./.venv/lib/python3.12/site-packages (from requests>=2.22.0->grss==4.3.8) (3.10)
Requirement already satisfied: urllib3<3,>=1.21.1 in ./.venv/lib/python3.12/site-packages (from requests>=2.22.0->grss==4.3.8) (2.3.0)
Requirement already satisfied: certifi>=2017.4.17 in ./.venv/lib/python3.12/site-packages (from requests>=2.22.0->grss==4.3.8) (2025.1.31)
Requirement already satisfied: soupsieve>1.2 in ./.venv/lib/python3.12/site-packages (from beautifulsoup4>=4.8->astroquery>=0.4.0->grss==4.3.8) (2.6)
Requirement already satisfied: typing-extensions>=4.0.0 in ./.venv/lib/python3.12/site-packages (from beautifulsoup4>=4.8->astroquery>=0.4.0->grss==4.3.8) (4.12.2)
Requirement already satisfied: six>=1.9 in ./.venv/lib/python3.12/site-packages (from html5lib>=0.999->astroquery>=0.4.0->grss==4.3.8) (1.17.0)
Requirement already satisfied: webencodings in ./.venv/lib/python3.12/site-packages (from html5lib>=0.999->astroquery>=0.4.0->grss==4.3.8) (0.5.1)
Collecting SecretStorage>=3.2 (from keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for SecretStorage>=3.2 from https://files.pythonhosted.org/packages/54/24/b4293291fa1dd830f353d2cb163295742fa87f179fcc8a20a306a81978b7/SecretStorage-3.3.3-py3-none-any.whl.metadata
  Using cached SecretStorage-3.3.3-py3-none-any.whl.metadata (4.0 kB)
Collecting jeepney>=0.4.2 (from keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for jeepney>=0.4.2 from https://files.pythonhosted.org/packages/b2/a3/e137168c9c44d18eff0376253da9f1e9234d0239e0ee230d2fee6cea8e55/jeepney-0.9.0-py3-none-any.whl.metadata
  Using cached jeepney-0.9.0-py3-none-any.whl.metadata (1.2 kB)
Collecting jaraco.classes (from keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for jaraco.classes from https://files.pythonhosted.org/packages/7f/66/b15ce62552d84bbfcec9a4873ab79d993a1dd4edb922cbfccae192bd5b5f/jaraco.classes-3.4.0-py3-none-any.whl.metadata
  Using cached jaraco.classes-3.4.0-py3-none-any.whl.metadata (2.6 kB)
Collecting jaraco.functools (from keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for jaraco.functools from https://files.pythonhosted.org/packages/9f/4f/24b319316142c44283d7540e76c7b5a6dbd5db623abd86bb7b3491c21018/jaraco.functools-4.1.0-py3-none-any.whl.metadata
  Using cached jaraco.functools-4.1.0-py3-none-any.whl.metadata (2.9 kB)
Collecting jaraco.context (from keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for jaraco.context from https://files.pythonhosted.org/packages/ff/db/0c52c4cf5e4bd9f5d7135ec7669a3a767af21b3a308e1ed3674881e52b62/jaraco.context-6.0.1-py3-none-any.whl.metadata
  Using cached jaraco.context-6.0.1-py3-none-any.whl.metadata (4.1 kB)
Collecting cryptography>=2.0 (from SecretStorage>=3.2->keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for cryptography>=2.0 from https://files.pythonhosted.org/packages/89/33/c1cf182c152e1d262cac56850939530c05ca6c8d149aa0dcee490b417e99/cryptography-44.0.2-cp39-abi3-manylinux_2_34_x86_64.whl.metadata
  Using cached cryptography-44.0.2-cp39-abi3-manylinux_2_34_x86_64.whl.metadata (5.7 kB)
Collecting more-itertools (from jaraco.classes->keyring>=15.0->astroquery>=0.4.0->grss==4.3.8)
  Obtaining dependency information for more-itertools from https://files.pythonhosted.org/packages/23/62/0fe302c6d1be1c777cab0616e6302478251dfbf9055ad426f5d0def75c89/more_itertools-10.6.0-py3-none-any.whl.metadata
  Using cached more_itertools-10.6.0-py3-none-any.whl.metadata (37 kB)
Requirement already satisfied: cffi>=1.12 in ./.venv/lib/python3.12/site-packages (from cryptography>=2.0->SecretStorage>=3.2->keyring>=15.0->astroquery>=0.4.0->grss==4.3.8) (1.17.1)
Requirement already satisfied: pycparser in ./.venv/lib/python3.12/site-packages (from cffi>=1.12->cryptography>=2.0->SecretStorage>=3.2->keyring>=15.0->astroquery>=0.4.0->grss==4.3.8) (2.22)
Using cached astropy-7.0.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (10.4 MB)
Using cached astroquery-0.4.9.post1-py3-none-any.whl (11.1 MB)
Using cached basemap-1.4.1-cp312-cp312-manylinux1_x86_64.whl (950 kB)
Using cached healpy-1.18.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (8.9 MB)
Using cached lxml-5.3.1-cp312-cp312-manylinux_2_28_x86_64.whl (5.0 MB)
Using cached matplotlib-3.8.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (11.6 MB)
Using cached numba-0.61.0-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl (3.9 MB)
Using cached numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (18.0 MB)
Using cached pandas-2.2.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (12.7 MB)
Using cached tqdm-4.67.1-py3-none-any.whl (78 kB)
Using cached astropy_iers_data-0.2025.3.10.0.29.26-py3-none-any.whl (1.9 MB)
Using cached basemap_data-1.3.2-py2.py3-none-any.whl (30.5 MB)
Using cached contourpy-1.3.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (323 kB)
Using cached cycler-0.12.1-py3-none-any.whl (8.3 kB)
Using cached fonttools-4.56.0-cp312-cp312-manylinux_2_5_x86_64.manylinux1_x86_64.manylinux_2_17_x86_64.manylinux2014_x86_64.whl (4.9 MB)
Using cached html5lib-1.1-py2.py3-none-any.whl (112 kB)
Using cached keyring-25.6.0-py3-none-any.whl (39 kB)
Using cached kiwisolver-1.4.8-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (1.5 MB)
Using cached llvmlite-0.44.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (42.4 MB)
Using cached packaging-23.2-py3-none-any.whl (53 kB)
Using cached pillow-11.1.0-cp312-cp312-manylinux_2_28_x86_64.whl (4.5 MB)
Using cached pyerfa-2.0.1.5-cp39-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (738 kB)
Using cached pyparsing-3.2.1-py3-none-any.whl (107 kB)
Using cached pyproj-3.6.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (8.7 MB)
Using cached pyshp-2.3.1-py2.py3-none-any.whl (46 kB)
Using cached pytz-2025.1-py2.py3-none-any.whl (507 kB)
Using cached pyvo-1.6.1-py3-none-any.whl (997 kB)
[…]
