### 2025-03-13 rahil-makadia/grss#90 @athulpg007
Community guidelines missing
Please add some guidelines for third parties wishing to
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

Example references:

- https://github.com/astropy/astropy/blob/main/CONTRIBUTING.md
- https://github.com/poliastro/poliastro/blob/main/CONTRIBUTING.md
