### 2025-03-13 rahil-makadia/grss#89 @athulpg007
Because of this, I could not run the test scripts locally.
```
(.venv) athul@athul-Lenovo-YOGA-720-13IKB:~/Documents/grss/tests/python$ ./run_tests.sh prop
[NbConvertApp] Converting notebook apophis.ipynb to script
[NbConvertApp] Writing 1953 bytes to apophis.py
[NbConvertApp] Converting notebook didymos.ipynb to script
[NbConvertApp] Writing 2037 bytes to didymos.py
[NbConvertApp] Converting notebook icarus.ipynb to script
[NbConvertApp] Writing 1759 bytes to icarus.py
[NbConvertApp] Converting notebook moshup.ipynb to script
[NbConvertApp] Writing 1763 bytes to moshup.py
Traceback (most recent call last):
  File "/home/athul/Documents/grss/tests/python/prop/apophis.py", line 9, in <module>
    import grss
ModuleNotFoundError: No module named 'grss'

```
