### 2025-03-13 721bc3f
added contributing file; placed cspice retrieval back into initialize.sh for source installs
