### 2025-03-13 rahil-makadia/grss#89 @rahil-makadia
Hi @athulpg007, thanks for flagging this down. I removed cspice as a dependency for the source code and while ago but it is still needed for some testing. I put back the cspice retrieval in the initialize script. If you repeat the process off the dev branch, it should work now. If you still get a pybind11 error, that might be system-specific though...
