### 2025-03-13 #7861 @rahil-makadia
Thanks for the issues @athulpg007. I think I have pushed commits to the `dev` branch that address both of them. Let me know if you still have problems with the installation or need the contributing file to have more specific content
