### 2025-03-15 rahil-makadia/grss#89 @athulpg007
Thank you, with the fix I was able to install grss from source and run the tests in the `dev` branch.
This issue can be closed when #91 is merged.
