### 2025-03-15 rahil-makadia/grss#92 @athulpg007
Consider adding some markdown comments before the cells in one example notebook
I stepped through the [example](https://rahil-makadia.github.io/grss/tests/python/prop/apophis.html) in the docs. Please consider:

- adding some markdown text before the cells to help the user understand what the various cells do.
- adding a note on where are the reference values in `jpl = np.array([....])` obtained from?
- adding labels to indicate the orbits of Apophis and planets. Currently, the orbits have no labels, though it can be seen it's Apophis and the planets. It would be nicer if there are labels.
