### 2025-03-15 #7861 @athulpg007
I have reviewed the submission following the checklist. The paper is well-written, the code and docs are well-organized. The issues I raised above have been addressed. Thank you! @rahil-makadia opened one more minor issue, and a comment:

- https://github.com/rahil-makadia/grss/issues/92 [optional]

- https://doi.org/10.3847/PSJ/adbc88 currently does not resolve, but I assume it will resolve once the PSJ paper is published.
