### 2025-03-16 1704f1f
added contributing note to readme; updated text in examples
