### 2025-03-16 rahil-makadia/grss#92 @rahil-makadia
the latest commits update the notebooks and address most of these reccomendations. When I find the time, I might go ahead and add more notes but Since this was an optional issue, I'm just gonna go ahead and close it for now
