### 2025-03-21 #7861 @rahil-makadia
@warrickball do you have the potential issue, number and DOI for this paper? I received the proofs for the PSJ paper and need to update the JOSS citation there
