### 2025-03-21 #7861 @warrickball
@rahil-makadia — Not yet. These will be generated once I test run the publication pipeline after the reviews. I've never looked closely before but I think you can reverse engineer the DOI and page number by comparing GitHub issue numbers for [some recently published papers](https://joss.theoj.org/papers/published)... I'm not sure how the volume and issue numbers are generated, though.

@ysBach has kindly stepped up as an additional reviewer, and we can accept the article as long as at least two reviewers recommend we do so.
