### 2025-03-21 #7861 @augustfly
@rahil-makadia I can chime in a little re what IOP/PSJ wants. I suspect that what @warrickball says will be sufficient for the PSJ proofs -- assuming this review will complete in 2025 we know that the volume will be 10 (for Year = 2025), and "page" is 7861 and the DOI: https://doi.org/10.21105/joss.07861 -- which are as warrick said are based on the github issue number. You only won't know the "number" or issue number until the month it is published, where the issue number counts forward from the 1st month JOSS started, is currently 107, and keeps iterating...forever.  :smile:
