### 2025-03-21 #7861 @rahil-makadia
I think the DOI and the page number might be enough for the proofs. I'll also hold onto them for a week or two in case the second review also happens quickly. Thanks for both of your help!
