### 2025-03-23 #7861 @ysBach
I have completed my review of the paper and GRSS, following the provided checklist. (I had already looked into the documentations a few weeks ago when I was initially contacted, which made this review process relatively quick.) In my view, the code and documentation are already at a publication-ready stage. I also reviewed some parts of the source code to verify calculations that I am familiar with. Plus, the software has already been used in scientific publications, demonstrating its practical use cases.

@warrickball I, therefore, recommend this paper for publication.
