### 2025-03-24 #7861 @warrickball
I'm just doing some quick checks myself (currently downloading kernels in one of the examples) but we'll then start moving through [the post-review process](https://joss.readthedocs.io/en/latest/editing.html#post-review). The main things we'll need are

* a new, tagged release version of the software, presuming its changed since v4.3.6, and
* an archive (e.g. on Zenodo, FigShare or another similar service) of that release version.

I don't have any immediate editorial comments and I'll have a close look once we've added the archive metadata.
