### 2025-03-24 #7861 @warrickball
Thanks again @athulpg007 & @ysBach for your prompt reviews. @rahil-makadia, your paper is now accepted. As mentioned, please make a new release of GRSS (if there have been changes to the code), archive that release to e.g. Zenodo, FigShare, and let me know the archive link and code version number. I'll then add the metadata to this review.

In the meantime I'll start having a close look at the text, though I won't finish that today.
