### 2025-03-25 528ef63
metadata cleanup; joss paper update; J2 smoothing partials update
