+++ joss/paper.bib
@@ -31,7 +31,8 @@ author    = {Rahil Makadia and Davide Farnocchia and Steven R. Chesley and Siegf
 journal   = {The Planetary Science Journal},
 title     = {{Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Planetary Defense}},
 year      = {2025},
-volume    = {Accepted},
+volume    = {6},
+doi       = {10.3847/PSJ/adbc88},
 publisher = {The American Astronomical Society},
 }
 
