### 2025-03-25 #7861 @rahil-makadia
@warrickball I updated the text with the DOI for the PSJ paper. None of the actual text in the JOSS paper has changed
