### 2025-03-26 22ddf7e
Initial editorial tweaks for JOSS paper
