+++ joss/paper.md
@@ -3,10 +3,10 @@ title: 'Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Plan
 tags:
   - Python
   - C++
-  - Asteroids
-  - Comets
-  - Orbit Determination
-  - Orbit Propagation
+  - asteroids
+  - comets
+  - orbit determination
+  - orbit propagation
 authors:
   - name: Rahil Makadia
     corresponding: true
@@ -33,15 +33,15 @@ aas-journal: Planetary Science Journal
 
 # Statement of Need
 
-Modeling the motion of small solar system bodies is of utmost importance when looking at the problem in the context of planetary defense. Reliably computing the orbit of an asteroid or a comet from various observations, and then predicting its trajectory in the future is critical in assessing the associated Earth-impact hazard. The NASA Center for Near-Earth Object Studies at the Jet Propulsion Laboratory (JPL) has developed a suite of state-of-the-art tools over the course of decades for this specific purpose. However, these tools are not publicly available. With the expected increase in the number of Near-Earth Object observations as well as discoveries when new observatories such as the Vera C. Rubin Observatory come online [@Schwamb2023], there is a need for a publicly available library that can reliably perform both orbit determination and propagation for small bodies in the solar system. Such a publicly available library will enable community efforts in planetary defense research.
+Modeling the motion of small Solar System bodies is of utmost importance when looking at the problem in the context of planetary defense. Reliably computing the orbit of an asteroid or a comet from various observations, and then predicting its trajectory in the future is critical in assessing the associated Earth-impact hazard. The NASA Center for Near-Earth Object Studies at the Jet Propulsion Laboratory (JPL) has developed a suite of state-of-the-art tools over the course of decades for this specific purpose. However, these tools are not publicly available. With the expected increase in the number of Near-Earth Object observations as well as discoveries when new observatories such as the Vera C. Rubin Observatory come online [@Schwamb2023], there is a need for a publicly available library that can reliably perform both orbit determination and propagation for small bodies in the solar system. Such a publicly available library will enable community efforts in planetary defense research.
 
 # Summary
 
-In this paper, we present ``GRSS``, the Gauss-Radau Small-body Simulator, an open-source library for orbit determination and propagation of small bodies in the solar system. ``GRSS`` is an open-source software library with a C++11 foundation and a Python binding. The propagator is based on the ``IAS15`` algorithm, a 15<sup>th</sup> order integrator based on Gauss-Radau quadrature [@Rein2014]. Only the particles of interest are integrated within ``GRSS`` to reduce computational cost. The states for the planets and 16 largest main-belt asteroids are computed using memory-mapped JPL digital ephemeris kernels [@Park2021] as done in the ``ASSIST`` orbit propagator library [@Holman2023]. In addition to the propagator, the C++ portion of the library also has the ability to predict impacts and calculate close encounter circumstances using various formulations of the B-plane [@Kizner1961; @Opik1976; @Chodas1999; @Milani1999; @Farnocchia2019].
[…]
