+++ joss/paper.bib
@@ -8,7 +8,7 @@ volume = {3},
 number = {8},
 pages = {184},
 author = {Rahil Makadia and Sabina D. Raducan and Eugene G. Fahnestock and Siegfried Eggl},
-title = {Heliocentric Effects of the DART Mission on the (65803) Didymos Binary Asteroid System},
+title = {Heliocentric Effects of the DART Mission on the (65803) {Didymos} Binary Asteroid System},
 journal = {The Planetary Science Journal},
 abstract = {The Double Asteroid Redirect Test (DART) is NASA’s first kinetic impact–based asteroid deflection mission. The DART spacecraft will act as a projectile during a hypervelocity impact on Dimorphos, the secondary asteroid in the (65803) Didymos binary system, and alter its mutual orbital period. The initial momentum transfer between the DART spacecraft and Dimorphos is enhanced by the ejecta flung off the surface of Dimorphos. This exchange is characterized within the system by the momentum enhancement parameter, β, and on a heliocentric level by its counterpart, β ⊙. The relationship between β and the physical characteristics of Dimorphos is discussed here. A nominal set of Dimorphos physical parameters from the design reference asteroid and impact circumstances from the design reference mission are used to initialize the ejecta particles for dynamical propagation. The results of this propagation are translated into a gradual momentum transfer onto the Didymos system barycenter. A high-quality solar system propagator is then used to produce precise estimates of the post-DART encounters between Didymos and Earth by generating updated close approach maps. Results show that even for an unexpectedly high β ⊙, a collision between the Didymos system and Earth is practically excluded in the foreseeable future. A small but significant difference is found in modeling the overall momentum transfer when individual ejecta particles escape the Didymos system, as opposed to imparting the ejecta momentum as a single impulse at impact. This difference has implications for future asteroid deflection campaigns, especially when it is necessary to steer asteroids away from gravitational keyholes.}
 }
@@ -16,7 +16,7 @@ abstract = {The Double Asteroid Redirect Test (DART) is NASA’s first kinetic i
 @article{Makadia2024,
 author    = {Rahil Makadia and Steven R. Chesley and Davide Farnocchia and Shantanu P. Naidu and Damya Souami and Paolo Tanga and Kleomenis Tsiganis and Masatoshi Hirabayashi and Siegfried Eggl},
 journal   = {The Planetary Science Journal},
-title     = {{Measurability of the Heliocentric Momentum Enhancement from a Kinetic Impact: The Double Asteroid Redirection Test (DART) Mission}},
+title     = {Measurability of the Heliocentric Momentum Enhancement from a Kinetic Impact: The {Double Asteroid Redirection Test (DART)} Mission},
 year      = {2024},
 month     = feb,
 number    = {2},
@@ -29,7 +29,7 @@ publisher = {The American Astronomical Society},
 @article{Makadia2025,
 author    = {Rahil Makadia and Davide Farnocchia and Steven R. Chesley and Siegfried Eggl},
 journal   = {The Planetary Science Journal},
-title     = {{Gauss-Radau Small-body Simulator (GRSS): An Open-Source Library for Planetary Defense}},
+title     = {{Gauss-Radau Small-body Simulator (GRSS)}: An Open-Source Library for Planetary Defense},
 year      = {2025},
 volume    = {6},
 doi       = {10.3847/PSJ/adbc88},
@@ -51,7 +51,7 @@ doi = {10.1093/mnras/stu2164}
 
[…]
