### 2025-03-26 #7861 @warrickball
I've opened a PR with some very minor capitalisation and markdown tweaks. Once they're merged, please continue with creating a Zenodo/FigShare/etc archive. If you make a new release, let me know what the corresponding version number is so I can update the metadata.
