### 2025-03-26 #7861 @rahil-makadia
Apologies @warrickball I accidentally clicked the wrong button and closed the PR (I'm choosing to blame the late hour and not my own badness). I went to your fork and ended up pushing the changes to dev directly. The JOSS review PR I had open will be merged soon and I'll go ahead and create the archive from main after that
