### 2025-03-26 #7861 @rahil-makadia
The new version is 4.4.0 and the DOI for the archive is [10.5281/zenodo.15087273](https://doi.org/10.5281/zenodo.15087273)
