### 2025-03-27 #7861 @warrickball
Thanks for the metadata. I've just done a first test run of the publication pipeline. I'll check the output this evening (specifically of some of the metadata artefacts) and we should be able to wrap this up. You can have a quick look yourself and flag an errors you spot but it's not required.

I don't think we need to worry about the invalid DOI for the associated PSJ paper because that should resolve once the two papers are published. If there's an issue, we do have capacity to fix it after publication.
