### 2025-03-27 #7861 @warrickball
Thanks once again to @athulpg007 & @ysBach for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.

Congratulations @rahil-makadia, your paper has been published in JOSS! 🎉
